Add files via upload
These are the testing images I downloaded and tested with. This is not a good practice. Because I didn't have certainty whether the images were already part of the dataset or not. They probably weren't though. The bigger problem is making sure the image is properly taken. Looking at the data, I think the images should be close-ups of the leaves.